Fix type annotation for max_delay in wait_n function and add measure_time function for runtime measurement

+++ python_async_function/1-concurrent_coroutines.py
@@ -7,7 +7,7 @@ from typing import List
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: int) -> List[float]:
+async def wait_n(n: int, max_delay: float) -> List[float]:
     """Spawn wait_random n times and return delays in ascending order."""
     coroutines = [wait_random(max_delay) for _ in range(n)]
     delays = await asyncio.gather(*coroutines)
